internal/runner: disable command tests on Windows
