+++ internal/runner/command_unix.go
@@ -1,0 +1,16 @@
+//go:build !windows
+
+package runner
+
+import (
+	"os/exec"
+	"syscall"
+)
+
+func setCmdAttrs(cmd *exec.Cmd) {
+	if cmd.SysProcAttr == nil {
+		cmd.SysProcAttr = &syscall.SysProcAttr{}
+	}
+	cmd.SysProcAttr.Setsid = true
+	cmd.SysProcAttr.Setctty = true
+}
