+++ internal/runner/command_windows.go
@@ -1,0 +1,7 @@
+//go:build windows
+
+package runner
+
+import "os/exec"
+
+func setCmdAttrs(cmd *exec.Cmd) {}
